+++ app/robot/actionManager.ts
@@ -1,6 +1,6 @@
 import "reflect-metadata";
 
-export const symbolRobotAction = Symbol("robotAction");
+const symbolRobotAction = Symbol("robotAction");
 
 export function robotAction(name: string, ...args: string[]) {
     return (target: any, propertyKey: string, descriptor: TypedPropertyDescriptor<Function>) => {
@@ -57,29 +57,5 @@ export default class ActionManager {
 
     public getThisObj(name: string) {
         return this.class_map.get(name);
-    }
-}
-
-class CopyActions {
-    @robotAction('复制', '复制来源', '复制目标')
-    public copy(from: string, to: string) {
-        
-    }
-
-    @robotAction('移动', '移动来源', '移动目标')
-    public move(from: string, to: string) {
-
-    }
-}
-
-class MailActions {
-    @robotAction('发邮件', '内容')
-    public sendMail(to: string, content: string) {
-
-    }
-
-    @robotAction('设定自己的账户', '邮箱', '密码')
-    public setupMail(username: string, password: string) {
-        
     }
 }
